+++ src/OpenTelemetry/Logs/ILogger/OpenTelemetryLoggerOptions.cs
@@ -36,6 +36,22 @@ public class OpenTelemetryLoggerOptions
     public bool IncludeScopes { get; set; }
 
     /// <summary>
+    /// Gets or sets the optional version of the instrumentation library reported
+    /// for logs produced via the <see cref="Microsoft.Extensions.Logging.ILogger"/>
+    /// integration. Default value: <see langword="null"/>.
+    /// </summary>
+    public string? Version { get; set; }
+
+    /// <summary>
+    /// Gets or sets the optional schema URL of the instrumentation library reported
+    /// for logs produced via the <see cref="Microsoft.Extensions.Logging.ILogger"/>
+    /// integration. Default value: <see langword="null"/>.
+    /// </summary>
+#pragma warning disable CA1056 // Change the type of property from 'string' to 'System.Uri'
+    public string? SchemaUrl { get; set; }
+#pragma warning restore CA1056 // Change the type of property from 'string' to 'System.Uri'
+
+    /// <summary>
     /// Gets or sets a value indicating whether or not log state should be
     /// parsed into <see cref="LogRecord.Attributes"/> on generated <see
     /// cref="LogRecord"/>s. Default value: <see langword="false"/>.
@@ -131,5 +147,7 @@ public class OpenTelemetryLoggerOptions
         ParseStateValues = this.ParseStateValues,
         IncludeAttributes = this.IncludeAttributes,
         IncludeTraceState = this.IncludeTraceState,
+        Version = this.Version,
+        SchemaUrl = this.SchemaUrl,
     };
 }